requirements.txt
prefect
minio
pandas
pyarrow
faker
streamlit
plotly
python-dotenv
pyspark==3.5.0
pymongo
fastapi
uvicorn
requests